% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oddvibe.R
\name{oddvibe}
\alias{oddvibe}
\title{ODdVIBe: Outlier Detection Via Iterated Boosting}
\description{
ODdVIBe: Outlier Detection Via Iterated Boosting
}
\references{
https://www.researchgate.net/publication/227006463_Iterated_Boosting_for_Outlier_Detection

https://link.springer.com/chapter/10.1007%2F3-540-34416-0_23#page-1
}
